fix IR representation
read the RTL for the IR implementation and figured out what
the IR format really means. The lower two bits are used to select
a mode for interpreting the 6-bit command.

For now the lower two bits are hard-coded, but an assert is
included above it to note if we have to fix that.

The parsing is pretty minimal for the .tex file but I don't see
any reason to implement cases that don't exist, generally I just
check the assumption with an assert and if we hit that we can add
the functionality.