fully unroll the inner loops of the fast jtag routines
This saves about 20-25% programming time.

The next optimization is probably not to iterate through a list
in python, but pass the entire block of data as a variable length
erray with a length parameter so it's iterated in C.

At the moment, it takes as long to issue the JTAG command, as
it does to simply iterate through the list one step. :-/ So
you'd get about a 30-40% speedup I think by doing that.